CEDS-4331 Update build configuration and dependencies for Java 11

+++ test/unit/uk/gov/hmrc/exports/scheduler/SchedulerDateUtilSpec.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 HM Revenue & Customs
+ * Copyright 2023 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,7 +27,7 @@ import uk.gov.hmrc.exports.config.AppConfig
 class SchedulerDateUtilSpec extends UnitSpec {
 
   private val zone = ZoneOffset.UTC
-  private val clock: Clock = Clock.fixed(instant("2019-01-01T12:00:00").plusNanos((Math.random() * 1000).toInt), zone)
+  private val clock: Clock = Clock.fixed(instant("2019-01-01T12:00:00").plusNanos((Math.random() * 1000).toLong), zone)
   private val config: AppConfig = mock[AppConfig]
   private val util = new SchedulerDateUtil(config)
